change: split parts of main.rs into lib.rs
this allows us to call e.g. collect_stats() during integration_testing
